+++ src/Memory/Common/DebugViews.cs
@@ -7,7 +7,7 @@ namespace Rustic.Memory;
 
 /// <summary>Proxy class used for displaying a <see cref="IReadOnlyCollection{T}"/> in the debugger.</summary>
 /// <typeparam name="T"></typeparam>
-public class IReadOnlyCollectionDebugView<T>
+public sealed class IReadOnlyCollectionDebugView<T>
 {
     private readonly WeakReference<IReadOnlyCollection<T>> _ref;